Changed type int -> double

+++ Tools/ProgressWatcher/src/progresswatcher/FileWatcherWorker.java
@@ -125,7 +125,7 @@ public class FileWatcherWorker implements Runnable {
 			// Get Progress valued
 			int time = Integer.parseInt(matcher.group(1));
 			double card = Double.parseDouble(matcher.group(2));
-			int estTime = Integer.parseInt(matcher.group(3));
+			double estTime = Double.parseDouble(matcher.group(3));
 			
 			NumberFormat format = NumberFormat.getInstance(Locale.GERMAN);
